+++ src/client/type/MappedContractType.ts
@@ -1,4 +1,4 @@
-import { MsgExecuteContract } from "@terra-money/terra.js";
+import { MsgExecuteContract } from "@xpla/xpla.js";
 
 export class ContractMsg extends MsgExecuteContract {