+++ src/client/xpla/XplaSingerKey.ts
@@ -1,6 +1,6 @@
 import { Key, SimplePublicKey } from "@xpla/xpla.js";
 import { Signer } from "../../signer/Signer";
-import { keccak256 } from "@ethersproject/keccak256";
+import { keccak_256 } from '@noble/hashes/sha3';
 
 export class XplaSignerKey extends Key {
   constructor(private signer: Signer, publicKey: Uint8Array) {
@@ -8,7 +8,7 @@ export class XplaSignerKey extends Key {
   }
 
   async sign(payload: Buffer): Promise<Buffer> {
-    const digest = Buffer.from(keccak256(payload).substring(2), 'hex');
+    const digest = keccak_256(payload);
     return Buffer.from(await this.signer.sign(digest));
   }
 }