+++ src/type/BlockchainType.ts
@@ -3,8 +3,7 @@
  */
 // export type BlockchainType =
 //   | "ethereum"
-//   | "terra"
 //   | "xpla"
 //   | "polygon"
-export const BLOCKCHAINS = ["ethereum", "terra", "xpla", "polygon"] as const;
+export const BLOCKCHAINS = ["ethereum", "xpla", "polygon"] as const;
 export type BlockchainType = typeof BLOCKCHAINS[number];