+++ CHANGELOG.md
@@ -1,10 +1,10 @@
 # Changelog
 ## [1.0.0] - 2023-03-02
 ### Added
-- Support for Ethereum(Polygon)
+- ([#CPBLOC-290](https://jira.com2us.com/jira/browse/CPBLOC-290)) Support for Ethereum(Polygon)
 
 ### Changed
-- Update client interface
+- ([#CPBLOC-290](https://jira.com2us.com/jira/browse/CPBLOC-290)) Update client interface
 - Bump `@xpla/xpla.js` from 0.2.0 to 0.3.1
 
 ## [0.1.0] - 2022-08-22